pyproject.toml
[project]
name = "bolas-locas"
version = "0.1.0"
description = ""
authors = [
    {name = "duplicandobitcoins", email = "duplicandobitcoins@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "pydantic (>=2.10.6,<3.0.0)"
]

packages = [{ include = "bolas_locas" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
